Move tests to pipeline tests
